Changes
So you were basically missing everything you needed to do on this page, you just changed the color scheme. For this part in the book we were supposed to make each of the letters at the top act as links to the sections on the page. Clicking "A" takes you to the start of the "A" section on the page, "B" takes you to the "B" section, etc. Also each section needed the word "Top" below them to take you back to the top of the page.
These are all done by "id" attributes. "<h2 id="startA">A</h2>" means that the letter "A" in the h2 tag has an id named "startA" attached to it. So at the top, the A there is changed to a link, "<a href="#startA">A</a>", that takes you to the A with the id. 
I probably made that confusing, too. To make it simpler, you're basically just putting hashtags like you do on Twitter or Instagram, just that in HTML the hashtags are in the code only.